separate the insertNewsletter functions form the newsletterData module
