+++ apps-rendering/src/server/newsletterData.ts
@@ -1,9 +1,5 @@
-import { Newsletter, Item } from 'item';
+import { Newsletter } from 'item';
 import fetch from 'node-fetch';
-import { Result, ResultKind } from '@guardian/types';
-import { BodyElement } from 'bodyElement';
-import { ElementKind } from 'bodyElementKind';
-import { ArticleDesign } from '@guardian/libs';
 import { Content } from '@guardian/content-api-models/v1/content';
 import { RenderingRequest } from '@guardian/apps-rendering-api-models/renderingRequest';
 
@@ -73,54 +69,6 @@ const getNewsletterData = async (
 	const data = await loadData();
 	const dataForId = data.find((_) => _.identityName === id);
 	return dataForId ? newsletterResponseToNewsletter(dataForId) : undefined;
-};
-
-const insertNewsletterIntoStandard = (
-	body: Array<Result<string, BodyElement>>,
-	newsletter: Newsletter,
-): Array<Result<string, BodyElement>> => {
-	const element: BodyElement = {
-		kind: ElementKind.NewsletterSignUp,
-		...newsletter,
-	};
-
-	body.unshift({
-		kind: ResultKind.Ok,
-		value: element,
-	});
-	return body;
-};
-
-export const insertNewsletterIntoItem = (
-	item: Item,
-	promotedNewsletter?: Newsletter,
-): Item => {
-	if (!promotedNewsletter) {
-		return item;
-	}
-
-	switch (item.design) {
-		case ArticleDesign.Standard:
-		case ArticleDesign.Gallery:
-		case ArticleDesign.Audio:
-		case ArticleDesign.Video:
-		case ArticleDesign.Review:
-		case ArticleDesign.Analysis:
-		case ArticleDesign.Comment:
-		case ArticleDesign.Feature:
-		case ArticleDesign.Recipe:
-		case ArticleDesign.MatchReport:
-		case ArticleDesign.Interview:
-		case ArticleDesign.Editorial:
-		case ArticleDesign.Obituary:
-			item.body = insertNewsletterIntoStandard(
-				item.body,
-				promotedNewsletter,
-			);
-			break;
-	}
-
-	return item;
 };
 
 export const providePromotedNewsletter = async (
